requirements.txt
pyserial
tensorflow
matplotlib
sklearn
keyboard
pygatt
pexpect
